[BEAM-7427] Fix JmsCheckpointMark Avro Encoding
